// src/store/modelStore/action.enum.ts
enum actionTypes {
  GET_MODELS_SUCCESS = "GET_MODELS_SUCCESS",
  GET_MODELS_PENDING = "GET_MODELS_PENDING",
  GET_MODELS_FAILED = "GET_MODELS_FAILED",

  ADD_MODEL_SUCCESS = "ADD_MODEL_SUCCESS",
  ADD_MODEL_FAILED = "ADD_MODEL_FAILED",
  ADD_MODEL_PENDING = "ADD_MODEL_PENDING",

  DELETE_MODEL_SUCCESS = "DELETE_MODEL_SUCCESS",
  DELETE_MODEL_FAILED = "DELETE_MODEL_FAILED",
  DELETE_MODEL_PENDING = "DELETE_MODEL_PENDING",
}

export default actionTypes;